# public/api/old_swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "version": "1.0.0",
    "title": "Modasti Backend API",
    "termsOfService": "",
    "contact": {
      "email": "biz@riseapps.co"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "stage-api.modasti.net",
  "basePath": "/api",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/itemDetails": {
      "post": {
        "tags": [
          "api"
        ],
        "summary": "item details",
        "description": "",
        "operationId": "itemDetails",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Item Details",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "itemId": {
                  "type": "number",
                  "required": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Item&Similar"
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/switchLike": {
      "post": {
        "tags": [
          "api"
        ],
        "summary": "switch like",
        "description": "",
        "operationId": "switchLike",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Switch Like",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "objId": {
                  "type": "number",
                  "required": true
                },
                "targetObject": {
                  "type": "string",
                  "enum": [
                    "item",
                    "set"
                  ],
                  "required": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "string",
                  "enum": [
                    "Like removed",
                    "Like added"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/getLikedItems": {
      "post": {
        "tags": [
          "api"
        ],
        "summary": "get Liked Items",
        "description": "",
        "operationId": "getLikedItems",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Liked Items",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "number"
                },
                "offset": {
                  "type": "number"
                },
                "limit": {
                  "type": "number"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "items": {
                  "$ref": "#/definitions/Items"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_token",
      "in": "query"
    }
  },
  "definitions": {
    "Items": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/Item"
      }
    },
    "Item": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "title_en": {
          "type": "string"
        },
        "price": {
          "type": "string"
        },
        "currency": {
          "type": "string"
        },
        "user_currency": {
          "type": "string"
        },
        "photos": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "table_id": {
                "type": "integer"
              },
              "photo_name": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "Item&Similar": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "title_en": {
          "type": "string"
        },
        "price": {
          "type": "string"
        },
        "currency": {
          "type": "string"
        },
        "user_currency": {
          "type": "string"
        },
        "photos": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "table_id": {
                "type": "integer"
              },
              "photo_name": {
                "type": "string"
              }
            }
          }
        },
        "similar": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Item"
          }
        }
      }
    },
...
  }
}